refactor(runtime,serverless): move request/response
